# api/merged-openapi.yaml
openapi: 3.0.3
info:
  title: Artemis API
  version: 0.0.1
  description: Artemis API
servers: []
paths:
...
  /user-management/api/local-student/register:
    post:
      tags:
        - UserManagement
      summary: registerLocalStudent
      description: registerLocalStudent endpoint to register a new local student
      operationId: registerLocalStudent
      security: []
      requestBody:
        description: Local student registration details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterLocalStudentRequest'
      responses:
        '200':
          description: Student registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterLocalStudentResponse'
  '/user-management/api/local-student/{localStudentId}':
    parameters:
      - name: localStudentId
        in: path
        description: ID of the local student
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - UserManagement
      summary: getLocalStudentById
      operationId: getLocalStudentById
      description: Retrieve local student details by their ID
      security:
        - keycloak_auth:
            - USER
            - TUTOR
            - ADMIN
      responses:
        '200':
          description: Local student details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalStudentProfile'
        '400':
          description: Bad request (invalid UUID format)
        '401':
          description: Unauthorized
        '404':
          description: Local student not found
    put:
      tags:
        - UserManagement
      summary: updateLocalStudentById
      operationId: updateLocalStudentById
      description: Update local student details by their ID
      security:
        - keycloak_auth:
            - USER
            - TUTOR
            - ADMIN
      requestBody:
        description: Local student details to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocalStudentProfileEditRequest'
      responses:
        '200':
          description: Local student details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalStudentProfile'
        '400':
          description: Bad request (invalid data)
        '401':
          description: Unauthorized
        '404':
          description: Local student not found
  '/user-management/api/local-student/{localStudentId}/profile-details':
    parameters:
      - name: localStudentId
        in: path
        description: ID of the local student
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - UserManagement
      summary: Complete local student profile details
      description: >-
        Allows a registered and approved local student to complete their profile
        details.
      operationId: completeLocalStudentProfile
      security:
        - keycloak_auth:
            - TUTOR
            - ADMIN
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LocalStudentProfileDetailsRequest'
      responses:
        '200':
          description: Profile details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalStudentProfileDetailsResponse'
        '400':
          description: Bad request (invalid data)
        '401':
          description: Unauthorized
  /user-management/api/local-student/assigned-international-students:
    get:
      tags:
        - UserManagement
      summary: getAssignedInternationalStudentsForLocalStudent
      description: >-
        Retrieve a list of international students assigned to the currently
        authenticated local student
      operationId: getAssignedInternationalStudentsForLocalStudent
      security:
        - keycloak_auth:
            - USER
            - TUTOR
            - ADMIN
      responses:
        '200':
          description: A list of assigned international students
          content:
            application/json:
              schema:
                type: object
                properties:
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssignedInternationalStudent'
        '401':
          description: Unauthorized
        '404':
          description: Local student not found
  '/user-management/api/local-student/assign-international-student/{internationalStudentId}':
    post:
      tags:
        - UserManagement
      summary: assignInternationalStudentToLocalStudent
      description: >-
        Assign an international student to the currently authenticated local
        student
      operationId: assignInternationalStudentToLocalStudent
      security:
        - keycloak_auth:
            - USER
            - TUTOR
            - ADMIN
      parameters:
        - name: internationalStudentId
          in: path
          description: ID of the international student to be assigned
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: International student assigned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: International student assigned successfully
        '400':
          description: Bad request (invalid UUID format)
        '401':
          description: Unauthorized
        '404':
          description: Local or international student not found
  /user-management/api/international-student/register:
    post:
      tags:
        - UserManagement
      summary: registerInternationalStudent
      description: Register a new international student
      operationId: registerInternationalStudent
      security: []
      requestBody:
        description: International student registration details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInternationalStudentRequest'
      responses:
        '200':
          description: Student registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterInternationalStudentResponse'
  '/user-management/api/international-student/{internationalStudentId}/profile-details':
    parameters:
      - name: internationalStudentId
        in: path
        description: ID of the international student
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - UserManagement
      summary: Complete international student profile details
      description: >-
        Allows a registered and approved international student to complete their
        profile details.
      operationId: completeInternationalStudentProfile
      security:
        - keycloak_auth:
            - TUTOR
            - ADMIN
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InternationalStudentProfileDetailsRequest'
      responses:
        '200':
          description: Profile details updated successfully
          content:
            application/json:
              schema:
                $ref: >-
                  #/components/schemas/InternationalStudentProfileDetailsResponse
        '400':
          description: Bad request (invalid data)
        '401':
          description: Unauthorized
  '/user-management/api/international-student/{internationalStudentId}':
    parameters:
      - name: internationalStudentId
        in: path
        description: ID of the international student
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - UserManagement
      summary: getInternationalStudentById
      operationId: getInternationalStudentById
      description: Retrieve international student details by their ID
      security:
        - keycloak_auth:
            - USER
            - TUTOR
            - ADMIN
      responses:
        '200':
          description: International student details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternationalStudentProfile'
        '400':
          description: Bad request (invalid UUID format)
        '401':
          description: Unauthorized
        '404':
          description: International student not found
    put:
      tags:
        - UserManagement
      summary: updateInternationalStudentById
      operationId: updateInternationalStudentById
      description: Update international student details by their ID
      security:
        - keycloak_auth:
            - USER
            - TUTOR
            - ADMIN
      requestBody:
        description: International student details to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternationalStudentProfileEditRequest'
      responses:
        '200':
          description: International student details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternationalStudentProfile'
        '400':
          description: Bad request (invalid data)
        '401':
          description: Unauthorized
        '404':
          description: International student not found
  /user-management/api/international-student/anonymous:
    get:
      tags:
        - UserManagement
      summary: getAllInternationalStudentsAnonymous
      description: Retrieve a list of all registered international students
      operationId: getAllInternationalStudentsAnonymous
      parameters:
        - name: page
          in: query
          description: Page number for pagination (default is 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Number of records per page for pagination (default is 20)
          required: false
          schema:
            type: integer
            default: 20
        - name: semesterId
          in: query
          description: Optional semester ID to filter students by semester
          required: false
          schema:
            type: string
            format: uuid
        - name: facultyId
          in: query
          description: Optional faculty ID to filter students by faculty
          required: false
          schema:
            type: string
            format: uuid
        - name: countryCode
          in: query
          description: Optional country code to filter students by country
          required: false
          schema:
            type: string
        - name: containAssigned
          in: query
          description: >-
            Whether to include students already assigned to local students
            (default is false)
          required: false
          schema:
            type: boolean
            default: false
      security:
        - keycloak_auth:
            - TUTOR
            - ADMIN
      responses:
        '200':
          description: A list of international students
          content:
            application/json:
              schema:
                type: object
                properties:
                  pageable:
                    $ref: '#/components/schemas/PageableResponse'
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/InternationalStudentAnonymous'
        '401':
          description: Unauthorized
...
components:
  schemas:
...
    GenderEnum:
      type: string
      enum:
        - MALE
        - FEMALE
        - OTHER
    PageableResponse:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        isLast:
          type: boolean
    StudentBase:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        bio:
          type: string
        favouriteActivities:
          type: array
          items:
            type: string
        profilePicturePath:
          type: string
        emailMarketingChecked:
          type: boolean
    ResponseStudentDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        facultyId:
          type: string
          format: uuid
        gender:
          $ref: '#/components/schemas/GenderEnum'
...
    StudentProfileBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        bio:
          type: string
        gender:
          $ref: '#/components/schemas/GenderEnum'
        dateOfBirth:
          type: string
          format: date
        facultyId:
          type: string
          format: uuid
        emailMarketingChecked:
          type: boolean
        termsAndConditionsChecked:
          type: boolean
        favouriteActivities:
          type: array
          items:
            type: string
    LocalStudentProfile:
      allOf:
        - $ref: '#/components/schemas/StudentProfileBase'
        - type: object
          properties:
            assignedInternationalStudents:
              type: array
              items:
                $ref: '#/components/schemas/AssignedInternationalStudent'
            assignedStudentsCapacity:
              type: integer
    InternationalStudentProfile:
      allOf:
        - $ref: '#/components/schemas/StudentProfileBase'
        - type: object
          properties:
            homeUniversity:
              type: string
            countryCode:
              type: string
            assignedBuddy:
              $ref: '#/components/schemas/ResponseStudentDetail'
            semestersIds:
              type: array
              items:
                type: string
                format: uuid
    StudentEditBaseRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        bio:
          type: string
        favouriteActivities:
          type: array
          items:
            type: string
        gender:
          $ref: '#/components/schemas/GenderEnum'
        facultyId:
          type: string
          format: uuid
        emailMarketingChecked:
          type: boolean
        dateOfBirth:
          type: string
          format: date
    LocalStudentProfileEditRequest:
      allOf:
        - $ref: '#/components/schemas/StudentEditBaseRequest'
    InternationalStudentProfileEditRequest:
      allOf:
        - $ref: '#/components/schemas/StudentEditBaseRequest'
        - type: object
          properties:
            homeUniversity:
              type: string
            countryCode:
              type: string
            accommodation:
              type: string
    InternationalStudentAnonymous:
      type: object
      properties:
        id:
          type: string
          format: uuid
        countryCode:
          type: string
        gender:
          $ref: '#/components/schemas/GenderEnum'
        homeUniversity:
          type: string
        facultyId:
          type: string
          format: uuid
        semesterIds:
          type: array
          items:
            type: string
            format: uuid
        bio:
          type: string
        isAssigned:
          type: boolean
    AssignedInternationalStudent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        countryCode:
          type: string
        gender:
          $ref: '#/components/schemas/GenderEnum'
        email:
          type: string
        phoneNumber:
          type: string
        homeUniversity:
          type: string
        facultyId:
          type: string
          format: uuid
        bio:
          type: string
        dateOfBirth:
          type: string
          format: date
    RegisterLocalStudentRequest:
      allOf:
        - $ref: '#/components/schemas/StudentBase'
        - type: object
          properties:
            facultyId:
              type: string
            dateOfBirth:
              type: string
              format: date
            gender:
              $ref: '#/components/schemas/GenderEnum'
            termsAndConditionsChecked:
              type: boolean
            password:
              type: string
          required:
            - firstName
            - lastName
            - email
            - password
            - dateOfBirth
            - phoneNumber
            - gender
            - termsAndConditionsChecked
            - facultyId
    LocalStudentProfileDetailsRequest:
      type: object
      properties:
        facultyId:
          type: string
          description: ID of the faculty
        description:
          type: string
          description: Short student bio
        emailMarketingChecked:
          type: boolean
          description: Consent to receive marketing emails
        profilePicture:
          type: string
          format: binary
          description: Profile picture file
      required:
        - facultyId
        - description
        - emailMarketingChecked
        - profilePicture
    RegisterInternationalStudentRequest:
      allOf:
        - $ref: '#/components/schemas/StudentBase'
        - type: object
          properties:
            countryISO:
              type: string
              pattern: '^[A-Z]{2}$'
            homeUniversity:
              type: string
            facultyId:
              type: string
            dateOfBirth:
              type: string
              format: date
            gender:
              $ref: '#/components/schemas/GenderEnum'
            termsAndConditionsChecked:
              type: boolean
            password:
              type: string
          required:
            - firstName
            - lastName
            - email
            - password
            - dateOfBirth
            - phoneNumber
            - gender
            - termsAndConditionsChecked
    InternationalStudentProfileDetailsRequest:
      type: object
      properties:
        homeUniversity:
          type: string
          description: Name of the home university
        accommodation:
          type: string
          description: Accommodation details
        facultyId:
          type: string
          description: ID of the faculty
        description:
          type: string
          description: Short student bio
        emailMarketingChecked:
          type: boolean
          description: Consent to receive marketing emails
        profilePicture:
          type: string
          format: binary
          description: Profile picture file
      required:
        - facultyId
        - description
        - emailMarketingChecked
        - profilePicture
    RegisterLocalStudentResponse:
      type: object
      properties:
        result:
          type: string
          example: Student registered successfully
    RegisterInternationalStudentResponse:
      type: object
      properties:
        result:
          type: string
          example: Student registered successfully
    LocalStudentProfileDetailsResponse:
      type: object
      properties:
        result:
          type: string
          example: Profile details updated successfully
    InternationalStudentProfileDetailsResponse:
      type: object
      properties:
        result:
          type: string
          example: Profile details updated successfully
...
  securitySchemes:
    keycloak_auth:
      type: oauth2
      description: OAuth2 flow via Keycloak
      flows:
        authorizationCode:
          authorizationUrl: >-
            https://keycloak.artemis.jirka.test.esnhk.cz/realms/Artemis/protocol/openid-connect/auth
          tokenUrl: >-
            https://keycloak.artemis.jirka.test.esnhk.cz/realms/Artemis/protocol/openid-connect/token
          scopes:
            ADMIN: Full administrative access
            USER: Standard user access
            TUTOR: Tutor-level access
...